requirements.txt
t-tech-investments
pandas
streamlit
python-dotenv
